refactor: move votingNeuronSelectStore to proposlDetail.store
