+++ frontend/svelte/src/lib/components/proposal-detail/VotingCard/VotingCard.svelte
@@ -4,7 +4,7 @@
   import { onDestroy } from "svelte";
   import { registerVotes } from "../../../services/proposals.services";
   import { i18n } from "../../../stores/i18n";
-  import { votingNeuronSelectStore } from "../../../stores/proposals.store";
+  import { votingNeuronSelectStore } from "../../../stores/proposalDetail.store";
   import Card from "../../ui/Card.svelte";
   import VotingConfirmationToolbar from "./VotingConfirmationToolbar.svelte";
   import CastVoteCardNeuronSelect from "./VotingNeuronSelect.svelte";
