+++ frontend/svelte/src/lib/stores/proposalDetail.store.ts
@@ -1,0 +1,48 @@
+import type {
+  NeuronId,
+  NeuronInfo,
+  ProposalId,
+  ProposalInfo,
+} from "@dfinity/nns";
+import { derived, writable } from "svelte/store";
+
+export interface NeuronSelectionStore {
+  neurons: NeuronInfo[];
+  selectedIds: NeuronId[];
+}
+
+/**
+ * Contains available for voting neurons and their selection state
+ */
+const initNeuronSelectStore = () => {
+  const { subscribe, update, set } = writable<NeuronSelectionStore>({
+    neurons: [],
+    selectedIds: [],
+  });
+
+  return {
+    subscribe,
+
+    set(neurons: NeuronInfo[]) {
+      set({
+        neurons: [...neurons],
+        selectedIds: neurons.map(({ neuronId }) => neuronId),
+      });
+    },
+
+    reset() {
+      this.set([]);
+    },
+
+    toggleSelection(neuronId: NeuronId) {
+      update(({ neurons, selectedIds }) => ({
+        neurons,
+        selectedIds: selectedIds.includes(neuronId)
+          ? selectedIds.filter((id: NeuronId) => id !== neuronId)
+          : Array.from(new Set([...selectedIds, neuronId])),
+      }));
+    },
+  };
+};
+
+export const votingNeuronSelectStore = initNeuronSelectStore();
