+++ frontend/svelte/src/tests/lib/components/proposal-detail/VotingCard/VotingConfirmationToolbar.spec.ts
@@ -7,7 +7,7 @@ import { fireEvent } from "@testing-library/dom";
 import { render, waitFor } from "@testing-library/svelte";
 import VotingConfirmationToolbar from "../../../../../lib/components/proposal-detail/VotingCard/VotingConfirmationToolbar.svelte";
 import { E8S_PER_ICP } from "../../../../../lib/constants/icp.constants";
-import { votingNeuronSelectStore } from "../../../../../lib/stores/proposals.store";
+import { votingNeuronSelectStore } from "../../../../../lib/stores/proposalDetail.store";
 import { formatVotingPower } from "../../../../../lib/utils/proposals.utils";
 import { mockNeuron } from "../../../../mocks/neurons.mock";
